Remove use of scala.ScalaObject
ScalaObject has been removed in Scala 2.11 and it was an implementation
detail that other projects shouldn't rely on. It looks like Play was using
ScalaObject to locate scala-library.jar. Let's switch to scala.Predef for
that purpose. It is a class that is likely to stay across all future Scala
versions.